// eshop-web/src/main/resources/client/src/app/shared/connect-common/enums/shopping-export-type.enum.ts
export enum EXPORT_TYPE {
    NEW = 'new',
    LEGACY = 'legacy'
}

export enum ADD_ON_ARTICLE_TYPE {
    DEPOSIT_ARTICLE = 'depositArticle',
    VRG_ARTICLE = 'vrgArticle',
    VOC_ARTICLE = 'vocArticle',
    PFAND_ARTICLE = 'pfandArticle'
}

export enum ADD_ON_ARTICLE_PREFIX {
    DEPOSIT_ARTICLE = 'DEP',
    VRG_ARTICLE = 'VRG',
    VOC_ARTICLE = 'VOC',
    PFAND_ARTICLE = 'PFA'
}